src/requirements.txt
requests
beautifulsoup4
sentence_transformers
torch
faiss-cpu
redis
regex
mariadb
spacy
aiohttp
lxml